concat_type_list and filter_type_list now in header Kokkos_Utilities.hpp and namespace Kokkos::Impl.  Added compile time unit tests in TestTypeList.cpp.
is_default_memory_trait now in header Kokkos_MemoryTraits.hpp and
namespace Kokkos::Impl.

device_memory_space and python_view_type_impl now in namespace Kokkos::Impl.
