+++ core/unit_test/CMakeLists.txt
@@ -70,6 +70,7 @@ KOKKOS_ADD_EXECUTABLE(
   SOURCES
   TestDetectionIdiom.cpp
   TestInterOp.cpp
+  TestTypeList.cpp
 )
 
 foreach(Tag Threads;Serial;OpenMP;Cuda;HPX;OpenMPTarget;HIP;SYCL)
